requirements.txt
psycopg2
python-dotenv
pathlib
pyyaml